feat: added switch network button on main page, removed unnesesary component, updated README, added remaining time on usequizcontract for future use
Task #000017
