+++ hooks/useWalletConnection.ts
@@ -1,6 +1,7 @@
 import { useWeb3React } from "@web3-react/core"
 import { injected } from "@/components/wallet/connectors"
 import { useEffect } from "react"
+import { Network } from "@/types"
 
 export const useWalletConnection = () => {
   const { active, account, activate, deactivate, chainId, library } = useWeb3React()
@@ -33,6 +34,34 @@ export const useWalletConnection = () => {
     }
   }
 
+  const goerliTestnetNetwork: Network = {
+    chainId: `0x${Number(5).toString(16)}`,
+    chainName: "Goerli Testnet",
+    nativeCurrency: {
+      name: "ETH",
+      symbol: "ETH",
+      decimals: 18
+    },
+    rpcUrls: ["https://goerli.infura.io/v3/0aee341c49d14f07b86fac7af12b7240"],
+    blockExplorerUrls: ["https://goerli.etherscan.io/"]
+  };
+
+  const changeNetwork = async (): Promise<void> => {
+    try {
+      if (!window.ethereum) throw new Error("No crypto wallet detected");
+      await window.ethereum.request({
+        method: "wallet_addEthereumChain",
+        params: [
+          {
+            ...goerliTestnetNetwork,
+          },
+        ],
+      })
+    } catch (error) {
+      console.log(error)
+    }
+  }
+
   useEffect(() => {
     connectWalletOnPageLoad()
   }, [])
@@ -43,6 +72,7 @@ export const useWalletConnection = () => {
     chainId,
     connectWallet,
     disconnectWallet,
-    library
+    library,
+    changeNetwork
   }
 }
